requirements.txt
anthropic>=0.40.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.150.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
